Refactor BoxBorder struct and move it to its own file
