+++ components/box/box.go
@@ -6,32 +6,6 @@ import (
 	"github.com/Shresht7/go-cli-tools/helpers"
 )
 
-type BoxBorder struct {
-	top         string
-	topRight    string
-	right       string
-	bottomRight string
-	bottom      string
-	bottomLeft  string
-	left        string
-	topLeft     string
-}
-
-// BorderSingle is a BoxBorder with single line characters
-// ┌────────┐
-// │        │
-// └────────┘
-var BorderSingle = BoxBorder{
-	top:         "─",
-	topRight:    "┐",
-	right:       "│",
-	bottomRight: "┘",
-	bottom:      "─",
-	bottomLeft:  "└",
-	left:        "│",
-	topLeft:     "┌",
-}
-
 type Box struct {
 	content string
 	border  BoxBorder
